fix: change execute function name in Command
Execute function in constructor is now called handler and it's private.
this function will be called by a newly added 'execute' function which
is public. The point is we can now call the handler with or without
args. So execute function will be the same for both 'SingleCommand' and
'CommandWithArgs' but the the handler function will differ.